requirements.txt
PyYAML
umapi-client